store-locator-api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
bcrypt==5.0.0
certifi==2025.11.12
click==8.3.1
Django==5.2.8
fastapi==0.127.0
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
python-dotenv==1.2.1
SQLAlchemy==2.0.45
sqlparse==0.5.4
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
